+++ includes/libraries/culqi-php/lib/Culqi/Cargos.php
@@ -3,21 +3,21 @@ namespace Culqi;
 
 class Cargos extends Resource {
 
-    const URL_CARGOS = "/cargos/";
+    const URL_CARGOS = "/charges/";
 
     public function create($options = NULL)
     {
-        return $this->request("POST", Cargos::URL_CARGOS, $api_key = $this->culqi->api_key, $options);
+        return $this->request("POST", self::URL_CARGOS, $api_key = $this->culqi->api_key, $options);
     }
 
     public function getList($options = NULL)
     {
-        return $this->request("GET", Cargos::URL_CARGOS, $api_key = $this->culqi->api_key, $options);
+        return $this->request("GET", self::URL_CARGOS, $api_key = $this->culqi->api_key, $options);
     }
 
     public function get($id)
     {
-        return $this->request("GET", Cargos::URL_CARGOS . $id . "/", $api_key = $this->culqi->api_key);
+        return $this->request("GET", self::URL_CARGOS . $id . "/", $api_key = $this->culqi->api_key);
     }