+++ includes/libraries/culqi-php/lib/Culqi/Suscripciones.php
@@ -4,11 +4,11 @@ namespace Culqi;
 
 class Suscripciones extends Resource {
 
-    const URL_SUSCRIPCIONES = "/suscripciones/";
+    const URL_SUSCRIPCIONES = "/subscriptions/";
 
     public function create($options = NULL)
     {
-        return $this->request("POST", Suscripciones::URL_SUSCRIPCIONES, $api_key = $this->culqi->api_key, $options);
+        return $this->request("POST", self::URL_SUSCRIPCIONES, $api_key = $this->culqi->api_key, $options);
     }
     //
     // public function getList($options = NULL) {