+++ includes/libraries/culqi-php/lib/Culqi/Tokens.php
@@ -1,28 +1,20 @@
 <?php
 namespace Culqi;
 
-class Tokens extends Resource
-{
+class Tokens extends Resource {
+
     /**
      * @param string $id La ID del token a devolver.
      * @param array|string|null $opts
      *
      * @return Token
      */
-    public static function retrieve($id, $opts = null)
+
+    const URL_TOKENS = "/tokens/";
+
+    public function create($options = NULL)
     {
-        return self::_retrieve($id, $opts);
+        return $this->request("POST", Tokens::URL_TOKENS, $api_key = $this->culqi->api_key, $options);
     }
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Token Token creado
-     * ATENCIÓN: Solo para desarollo. Lo ideal es usar el CULQI.JS.
-     *
-     */
-    public static function create($params = null, $opts = null)
-    {
-        return self::_create($params, $opts);
-    }
+
 }